requirements.txt
flask
werkzeug
openai-whisper
requests
imageio-ffmpeg
soundfile
numpy
